requirements.txt
streamlit
google-generativeai
python-dotenv
diffusers
torch
transformers

